Refactor useUserViews hook
