+++ src/scripts/libraryMenu.js
@@ -2,7 +2,7 @@ import escapeHtml from 'escape-html';
 import Headroom from 'headroom.js';
 
 import { AppFeature } from 'constants/appFeature';
-import { getUserViewsQuery } from 'hooks/useUserViews';
+import { getUserViewsQuery } from 'hooks/api/useUserViews';
 import globalize from 'lib/globalize';
 import { ServerConnections } from 'lib/jellyfin-apiclient';
 import { EventType } from 'constants/eventType';
@@ -392,7 +392,7 @@ function onSidebarLinkClick() {
 
 function getUserViews(apiClient, userId) {
     return queryClient
-        .fetchQuery(getUserViewsQuery(toApi(apiClient), userId))
+        .fetchQuery(getUserViewsQuery(toApi(apiClient), { userId }))
         .then(function (result) {
             const items = result.Items;
             const list = [];
